moved add application to drawer
